Extract out into cli.rs
