+++ nym-vpnc/src/main.rs
@@ -4,7 +4,8 @@
 use std::path::{Path, PathBuf};
 
 use anyhow::{anyhow, Context};
-use clap::{Args, Parser, Subcommand};
+use clap::Parser;
+use cli::Command;
 use nym_gateway_directory::{EntryPoint, ExitPoint, NodeIdentity, Recipient};
 use nym_vpn_proto::{
     nym_vpnd_client::NymVpndClient, ConnectRequest, DisconnectRequest, ImportUserCredentialRequest,
@@ -13,120 +14,11 @@ use nym_vpn_proto::{
 use parity_tokio_ipc::Endpoint as IpcEndpoint;
 use tonic::transport::{Channel as TonicChannel, Endpoint as TonicEndpoint};
 
-#[derive(Parser)]
-#[clap(author = "Nymtech", version, about)]
-struct CliArgs {
-    /// Use HTTP instead of socket file for IPC with the daemon.
-    #[arg(long)]
-    http: bool,
+use crate::cli::ImportCredentialTypeEnum;
 
-    #[command(subcommand)]
-    command: Command,
-}
+mod cli;
 
-#[derive(Subcommand)]
-enum Command {
-    Connect(ConnectArgs),
-    Disconnect,
-    Status,
-    ImportCredential(ImportCredentialArgs),
-}
-
-#[derive(Args)]
-pub(crate) struct ConnectArgs {
-    #[command(flatten)]
-    pub(crate) entry: CliEntry,
-
-    #[command(flatten)]
-    pub(crate) exit: CliExit,
-
-    /// Disable routing all traffic through the nym TUN device. When the flag is set, the nym TUN
-    /// device will be created, but to route traffic through it you will need to do it manually,
-    /// e.g. ping -Itun0.
-    #[arg(long)]
-    pub(crate) disable_routing: bool,
-
-    /// Enable two-hop mixnet traffic. This means that traffic jumps directly from entry gateway to
-    /// exit gateway.
-    #[arg(long)]
-    pub(crate) enable_two_hop: bool,
-
-    /// Enable Poisson process rate limiting of outbound traffic.
-    #[arg(long)]
-    pub(crate) enable_poisson_rate: bool,
-
-    /// Disable constant rate background loop cover traffic.
-    #[arg(long)]
-    pub(crate) disable_background_cover_traffic: bool,
-
-    /// Enable credentials mode.
-    #[arg(long)]
-    pub(crate) enable_credentials_mode: bool,
-}
-
-#[derive(Args)]
-#[group(multiple = false)]
-pub(crate) struct CliEntry {
-    /// Mixnet public ID of the entry gateway.
-    #[clap(long, alias = "entry-id")]
-    pub(crate) entry_gateway_id: Option<String>,
-
-    /// Auto-select entry gateway by country ISO.
-    #[clap(long, alias = "entry-country")]
-    pub(crate) entry_gateway_country: Option<String>,
-
-    /// Auto-select entry gateway by latency
-    #[clap(long, alias = "entry-fastest")]
-    pub(crate) entry_gateway_low_latency: bool,
-
-    /// Auto-select entry gateway randomly.
-    #[clap(long, alias = "entry-random")]
-    pub(crate) entry_gateway_random: bool,
-}
-
-#[derive(Args)]
-#[group(multiple = false)]
-pub(crate) struct CliExit {
-    /// Mixnet recipient address.
-    #[clap(long, alias = "exit-address")]
-    pub(crate) exit_router_address: Option<String>,
-
-    /// Mixnet public ID of the exit gateway.
-    #[clap(long, alias = "exit-id")]
-    pub(crate) exit_gateway_id: Option<String>,
-
-    /// Auto-select exit gateway by country ISO.
-    #[clap(long, alias = "exit-country")]
-    pub(crate) exit_gateway_country: Option<String>,
-
-    /// Auto-select exit gateway randomly.
-    #[clap(long, alias = "exit-random")]
-    pub(crate) exit_gateway_random: bool,
-}
-
-#[derive(Args)]
-pub(crate) struct ImportCredentialArgs {
-    #[command(flatten)]
-    pub(crate) credential_type: ImportCredentialType,
-
-    // currently hidden as there exists only a single serialization standard
-    #[arg(long, hide = true)]
-    pub(crate) version: Option<u8>,
-}
-
-#[derive(Args, Clone)]
-#[group(required = true, multiple = false)]
-pub(crate) struct ImportCredentialType {
-    /// Credential encoded using base58.
-    #[arg(long)]
-    pub(crate) credential_data: Option<String>,
-
-    /// Path to the credential file.
-    #[arg(long)]
-    pub(crate) credential_path: Option<PathBuf>,
-}
-
-fn parse_entry_point(args: &ConnectArgs) -> anyhow::Result<Option<EntryPoint>> {
+fn parse_entry_point(args: &cli::ConnectArgs) -> anyhow::Result<Option<EntryPoint>> {
     if let Some(ref entry_gateway_id) = args.entry.entry_gateway_id {
         Ok(Some(EntryPoint::Gateway {
             identity: NodeIdentity::from_base58_string(entry_gateway_id.clone())
@@ -145,7 +37,7 @@ fn parse_entry_point(args: &ConnectArgs) -> anyhow::Result<Option<EntryPoint>> {
     }
 }
 
-fn parse_exit_point(args: &ConnectArgs) -> anyhow::Result<Option<ExitPoint>> {
+fn parse_exit_point(args: &cli::ConnectArgs) -> anyhow::Result<Option<ExitPoint>> {
     if let Some(ref exit_router_address) = args.exit.exit_router_address {
         Ok(Some(ExitPoint::Address {
             address: Recipient::try_from_base58_string(exit_router_address.clone())
@@ -171,25 +63,9 @@ fn parse_encoded_credential_data(raw: &str) -> bs58::decode::Result<Vec<u8>> {
     bs58::decode(raw).into_vec()
 }
 
-// Workaround until clap supports enums for ArgGroups
-pub(crate) enum ImportCredentialTypeEnum {
-    Path(PathBuf),
-    Data(String),
-}
-
-impl From<ImportCredentialType> for ImportCredentialTypeEnum {
-    fn from(ict: ImportCredentialType) -> Self {
-        match (ict.credential_data, ict.credential_path) {
-            (Some(data), None) => ImportCredentialTypeEnum::Data(data),
-            (None, Some(path)) => ImportCredentialTypeEnum::Path(path),
-            _ => unreachable!(),
-        }
-    }
-}
-
 #[tokio::main]
 async fn main() -> anyhow::Result<()> {
-    let args = CliArgs::parse();
+    let args = cli::CliArgs::parse();
     match args.command {
         Command::Connect(ref connect_args) => connect(&args, connect_args).await?,
         Command::Disconnect => disconnect(&args).await?,
@@ -216,7 +92,7 @@ fn default_endpoint() -> String {
     "http://[::1]:53181".to_string()
 }
 
-async fn get_client(args: &CliArgs) -> anyhow::Result<NymVpndClient<TonicChannel>> {
+async fn get_client(args: &cli::CliArgs) -> anyhow::Result<NymVpndClient<TonicChannel>> {
     if args.http {
         let endpoint = default_endpoint();
         let client = NymVpndClient::connect(endpoint.clone())
@@ -304,7 +180,7 @@ fn into_exit_point(exit: ExitPoint) -> nym_vpn_proto::ExitNode {
     }
 }
 
-async fn connect(args: &CliArgs, connect_args: &ConnectArgs) -> anyhow::Result<()> {
+async fn connect(args: &cli::CliArgs, connect_args: &cli::ConnectArgs) -> anyhow::Result<()> {
     // Setup connect arguments
     let entry = parse_entry_point(connect_args)?;
     let exit = parse_exit_point(connect_args)?;
@@ -320,7 +196,7 @@ async fn connect(args: &CliArgs, connect_args: &ConnectArgs) -> anyhow::Result<(
     Ok(())
 }
 
-async fn disconnect(args: &CliArgs) -> anyhow::Result<()> {
+async fn disconnect(args: &cli::CliArgs) -> anyhow::Result<()> {
     let mut client = get_client(args).await?;
     let request = tonic::Request::new(DisconnectRequest {});
     let response = client.vpn_disconnect(request).await?.into_inner();
@@ -328,7 +204,7 @@ async fn disconnect(args: &CliArgs) -> anyhow::Result<()> {
     Ok(())
 }
 
-async fn status(args: &CliArgs) -> anyhow::Result<()> {
+async fn status(args: &cli::CliArgs) -> anyhow::Result<()> {
     let mut client = get_client(args).await?;
     let request = tonic::Request::new(StatusRequest {});
     let response = client.vpn_status(request).await?.into_inner();
@@ -337,8 +213,8 @@ async fn status(args: &CliArgs) -> anyhow::Result<()> {
 }
 
 async fn import_credential(
-    args: &CliArgs,
-    import_args: &ImportCredentialArgs,
+    args: &cli::CliArgs,
+    import_args: &cli::ImportCredentialArgs,
 ) -> anyhow::Result<()> {
     let import_type: ImportCredentialTypeEnum = import_args.credential_type.clone().into();
     let raw_credential = match import_type {
